Remove pushd from components load flow (#29766)
## Summary & Motivation

This pushd is problematic for a couple of reasons:
- If you have code that depends on this working directory being a
certain value, it will not work (an example would be storing a relative
path that you want to use later)
- It assumes that the folder actually exists on disk, which is not true
for some of the change preview features we have been experimenting with.

This PR removes that pushd. We will need to document this and probably
touch up the changelog entry a bit, maybe with a before/after example?

## How I Tested These Changes
First seeing what 

## Changelog
Dagster components no longer change the working directory while they are
being loaded. This allows components to store relative paths and ensure
that they will still work when accessed outside of the component loading
codepath. This change may affect user-defined components that depend on
`Path.cwd()` or `os.getcwd()`. Instead, you should use a path relative
to the current source file when loading paths in a component, using the
`context.resolve_source_relative_path` method (see `resolve_dbt_project`
in `DbtProjectComponent` for an example)