Catalog v2 perf + loading improvements (#29816)
## Summary & Motivation
1. More memoization
2. Add shimmer rows when health data is still loading to avoid going
from showing shimmer rows to no rows / 1 row when we have partial data
3. Add an in memory cache for asset selection resolution similar to what
we had in the cloud version of the catalog view tiles.
4. Use a separate live data thread for fetching health data in
useAssetGraphSupplementaryData in order to avoid starving the rest of
the UI out making the catalog load much faster.
5. Use JSON.stringify on keys in LiveData to avoid
unsubscribing/resubscribing to the same set of keys when references
change.
6. Process live data updates immediately when cache data is available
during subscription to avoid an artificial delay on page load.

## How I Tested These Changes

App-proxy. Existing tests for LiveDataProvider