add latest run id to `AssetMaterializationHealthState` (#29802)
## Summary & Motivation
In the asset health UIs we want to show the latest failed run id for an
asset in a failed state. I figured we might as well just store the id of
the latest run to either successfully materialize or fail to materialize
the asset. So this adds that metadata. I'm not thrilled with the name
`latest_terminal_run_id`, i'm worried it implies that canceled or
deleted runs may be included, which is not necessarily the case (unless
they successfully materialized an asset before canceling/deleting)

## How I Tested These Changes
in https://github.com/dagster-io/internal/pull/15480