enable declarative automation lineage navigation (#29516)
## Summary & Motivation
One main component of observability for DA is providing clarity on why a
given asset did or did not materialize in a particular materialization.
Often this is due to some condition of one of its dependencies. However,
it's currently very difficult to navigate the lineage tree in the DA
evaluation context.


https://github.com/user-attachments/assets/1da3952a-4288-41ac-a40e-f837c3e30580



## How I Tested These Changes
BK, above

## Changelog
[ui] Added asset lineage navigation within the automation condition
evaluation tree.