add automation condition information on the global asset graph (#29583)
## Summary & Motivation
A very natural way to explore the automation conditions (and the
relationship between asset dependencies) is via the asset graph view.

One thing to note is that we are now pulling a good bit more information
for every node (sensors/schedule state). Might be worthwhile to
progressively pull in that information only if the facet is enabled.


https://github.com/user-attachments/assets/7be87d8c-5fc7-489e-a5a8-633a0bcdaec9



## How I Tested These Changes
BK

## Changelog
Added a asset graph node facet for viewing automation conditions and the
most recent evaluation