[ui] Update the behavior of "view lineage" on the runs view (#29894)
## Summary & Motivation

https://dagsterlabs.slack.com/archives/C03CCE471E0/p1746749043149529

Previously, we would include the downstreams of each key on the run when you clicked the "View lineage" button.

I found this unintuitive, as the "view list" button only shows the keys directly on the run. I also added parens to the wider expression to make it easier to add further clauses (e.g. (<big expression>) and status:degraded))

## How I Tested These Changes