[dg] in process mode (#29362)
in the case that we are in a `dg` context where we are targeting the
python environment we are already in, avoid a process hop and call the
target cli directly via `CliRunner`


Here is a diagram visualizing things

![image](https://github.com/user-attachments/assets/fbf81ac8-5ae7-47f5-b749-bd227e8a7dc5)


The main remaining issue currently that led to some tests being kept on
`uv_managed` is that the plugin system depends on import entry group
metadata that is only set when the module is properly installed. I think
we could add env var injection for plugins to deal with this issue.

## How I Tested These Changes

update the test util defaults to leverage this new mode