[dagster-airlift][migration 3/n] introduce externally managed runs to core system (#29810)
## Summary & Motivation
Introduces "externally managed asset runs" to the core system. When a
specific tag is set, the yielding of asset materialization planned and
asset materialization events will be stopped, with the expectation that
an observer system will perform the actual event yielding later.

This is used in Airlift to maintain clean separation between the
"observer run" - that produced by the monitoring job process - and any
"externally managed runs" - those triggered by airflow directly due to
proxying behavior.

## How I Tested These Changes
Added new tests for the functionality. Downstream tested with airlift.