[dagster-airlift][migration 1/n] explicit methods for external jobs, use them in airlift (#29706)
This PR has been used to merge the large set of upstack changes which
comprise a P0 implementation of jobs-based airlift.

## Summary & Motivation
Adds explicit methods for constructing "external jobs". These jobs have
an asset layer with only asset keys defined, but can still be properly
linked back/forth from the asset by the code location.

The upshot is that we're able to create job definitions for assets which
are materializable _outside_ of a specific job - because there's no
direct linkage between the asset and an op within the job. Previously we
could only accept asset specs in the job creation code path.

## How I Tested These Changes
Added a new test, tested live in dagster dev.