[dagster-fivetran] Handle unavailable schema configs in fetch_fivetran_workspace_data (#29881)
## Summary & Motivation

Fixes AD-1290.

In some cases, the schema config doesn't exist, even if the connector is
connected and the schema status is ready. The Fivetran API request fails
with a 404 error in that case.

## How I Tested These Changes

New tests with BK.

## Changelog

[dagster-fivetran] An issue causing the Fivetran integration to fail
when the schema config does not exist for a connector has been fixed.