[ui] Improve image string display for code locations (#29930)
## Summary & Motivation

For code locations, display the full image string, and move the copy
behavior to an adjacent button.

This should make it easier to search for (and select) image names in the
code location history page in Dagster+, and fixes a text alignment issue
as well.

## How I Tested These Changes

View code location list, verify that the entire image name is rendered,
wrapping onto a new line as needed. Verify click-to-copy behavior.

View code location change history, verify same.

## Changelog

[ui] Make it easier to select and copy image names for code locations.