Update mechanism for re-execute from asset failure (#29890)
## Summary & Motivation

This changes the logic for determining which keys to execute to solve two issues:

We did not handle non-subsettable multi assets at all, so this process could have resulted in errors when generating the execution plan
When a blocking asset check fails, we should re-execute it (as this is the source of the run failure, we need to re-execute as part of the process), and we weren't doing that before
Potentially controversially, I have this re-executing both the blocking asset check AND the asset that the check targets, with the logic being that the code of the asset itself is more likely to be the issue than the code of the check. So the workflow you'd want would be to update the code of the asset, then execute "re-execute from asset failure", and have your asset re-executed.

Also moved some of the logic from using the event log scan to using the execution plan snapshot as suggest by Daniel, which might help perf somewhat.

## How I Tested These Changes