[gql] assetEventHistory resolver (#29845)
## Summary & Motivation
Adds a new resolver to `GrapheneAsset` to get a sorted list of all
Materialization, Failed to materialize, and Observation events, so that
we can display them in the correct order without doing multiple queries.
The endpoint allows you to pass a list of event types you want to
receive, so we can filter down to any subset of these events and still
have them in the correct order.

I think we can replace the `materializationHistory` endpoint with this
endpoint, but will do that in later once we migrate the UI over so that
it's all push safe. That would allow us to remove some of the
GrapheneTypes that are used only by the materializationHistory resolver
too.

more tests in https://github.com/dagster-io/internal/pull/15524 (with
failure events)