Use MurmurHash3 to hash keys for indexeddb caching (#29980)
## Summary & Motivation

Currently for large organizations our existing hash function just
JSON.stringifies the arguments to our graph layout algorithm. The issue
with this is that the arguments can be over 100MB when stringified which
leads to a lot of memory being used that can cause the browser to crash.
To get around this use MurmurHash3 to minimize memory usage.

## How I Tested These Changes

jest tests

App-proxy - navigate the app of an org with a large workspace and saw
that the memory crashing stopped occurring.

spot checked with console.logging that we were getting deterministic
hashes.
<img width="1089" alt="Screenshot 2025-05-14 at 12 42 42 PM"
src="https://github.com/user-attachments/assets/7fb35929-79af-440d-a91e-1a9c5bce8279"
/>