[dagster-fivetran] Support setting group names on per-asset basis thru translator (#29968)
## Summary

Right now, the Fivetran multi-asset is still built setting a group name
on the `multi_asset` itself. This errors if you try to return an
`AssetSpec` from the translator with a group name attached. This PR
moves the group name to specs, and tests that a translator can properly
alter the group name of an asset.