[dagster-airlift] Sensor changes for historical runs (#29895)
## Summary & Motivation
We've realized that allowing for "arbitrary history import" is a
trickier problem than we initially thought. Dagster currently
synonymizes increasing "storage id" in the event log with increasing
"event order", but Airlift V2 currently breaks this by potentially
inserting runs which occur in the "middle" of history - IE if I run the
airflow monitoring job on a time range which is not the most recently
processed one, I can insert runs in the middle of the timeline. This can
have weird operational affects on downstream event processors which will
pick up those runs in the middle of the timeline and kick off side
effects, like declarative automation and alerts

Instead of playing a game of "whack-a-mole" to make sure every callsite
properly handles "historical runs", we want to try to avoid the
problematic case in general by making it _impossible_ to insert runs in
the middle of the range. This has a few implications:
- Make the sensor hard fail if any kicked off monitoring job run fails;
so we can guarantee we only move "forward" in time.
- Since this makes it tricky to schedule a true "historical" run of the
monitoring job, start the sensor at the previous day when there is no
cursor. This still makes the initial "historical import" possible.

I'm making the explicit decision not to "hard fail" the monitoring job
for two reasons:
- It's not easy to query the monitoring job for a "later range" that
works in all cases; you can check the last run, but checking arbitrarily
far in the past gets tricky.
- There may be cases that come up where we _do_ want to rectify some
historical range, because we've decided that the downstream side effects
are acceptable. It's advantageous to be able to do that manually.


This is not a full solution. Eventually, we're going to need to figure
out how to separate "new to dagster" from "newly occurring", as this
will be broadly important for all federation-type integrations. We're
also going to want to track all the places that currently synonymize
"new to dagster" with "newly occurring" (basically all the things that
tail the event log).

Downstream tasks tracked here:
https://linear.app/dagster-labs/issue/BCOR-178/fully-native-historical-runs

## How I Tested These Changes
- New tests for failure cases.