Make externalJobSource retrievable without expensive queries (#29948)
## Summary & Motivation
This introduces "ref tags" to the JobRefSnap - a specific, select set of
tags that are always retrieved when constructing the "ref". For now I'm
just using this for "externalJobSource", but you can easily imagine
using this to resolve whether a job is an asset job for example (right
now that resolver seems to be pretty expensive).

## How I Tested These Changes
Ran UI directly locally