Fix timeout property data type in ClientConfig for Deltalake. (#28912)
## Summary & Motivation

The current implementation of ClientConfig expects the `timeout`
parameter to be an `int`.

However, this leads to errors when using the io manager, for instance,
to load an input:

```bash
dagster._core.errors.DagsterExecutionLoadInputError: Error occurred while loading input "xxx" of step "xxx": 

...

The above exception was caused by the following exception:OSError: Generic Config error: failed to parse "120" as Duration 
```

According to [this
discussion](https://github.com/delta-io/delta-rs/issues/2595) in the
deltalake community, the expected format for this parameter is a string
like "120s".

## How I Tested These Changes
Running locally with a modification to the DeltalakeDbClient to make the
proposed modification effective I was able to use the parameter without
the aforementioned error:

```python
class CustomDeltaLakeDbClient(DbClient):
    @staticmethod
    @contextmanager
    def connect(context, table_slice: TableSlice) -> Iterator[TableConnection]:
        
        ...

        storage_options = {
            **{k: str(v) for k, v in storage_options.items() if v is not None},
            **{k: str(v) for k, v in client_options.items() if v is not None},
        }
        # fix timeout setup
        if "timeout" in storage_options:
            storage_options["timeout"] += "s"

        ...        
```

## Changelog

[dagster-deltalake] Fix timeout property data type in ClientConfig to be
str instead of int

---------

Co-authored-by: Gareth Brickman <50636644+garethbrickman@users.noreply.github.com>