[dg] Make dagster-test a dg project (#30022)
## Summary & Motivation

This PR makes `dagster-test` itself a dg-compatible project. What this will allow us to do is have a pre-built project in the monorepo that doesn't require test-time installation. 

We can put lots of definitions and components in this project for tests cases, provided, of course, we keep it organized.

## How I Tested These Changes

`dg list defs` in directory

## Changelog

CHANGELOG