[dagster-dbt] Fix dbt cli invocation params for OpExecutionContext (#30003)
## Summary & Motivation

Fixes #29918.

The error was introduced in
https://github.com/dagster-io/dagster/pull/29650 when addressing this
[comment](https://github.com/dagster-io/dagster/pull/29650#discussion_r2064866321).

Context:

When creating an asset definition, it is allowed to use
`OpExecutionContext` as the type hint for the context, like

```python
@dbt_assets(manifest=manifest)
def my_dbt_assets(context: OpExecutionExecution, dbt: DbtCliResource):
    ...
```

Issue:

We removed the `suppress(DagsterInvalidPropertyError)` statement in
`get_updated_cli_invocation_params_for_context`, and used `assets_def =
context.assets_def if isinstance(context, AssetExecutionContext) else
None` instead. The problem is that the context may be typed as
`OpExecutionContext` even though the context is an instance of
`AssetExecutionContext`.

For that specific use case, we were not entering the if context and
assets_def is not None condition in the function, so the dbt cli params
were not properly returned.

Solution:

~~This PR brings back the `suppress(DagsterInvalidPropertyError)`
statement. We don't what type hint will be used for the context, and
using the suppress statement makes sure that Dagster doesn't yell when
accessing `context.asset_defs`.~~

This PR uses getattr to access context.asset_defs without using the
suppress statement and isinstance.


## How I Tested These Changes

Updated tests with BK

## Changelog

[dagster-dbt] An issue causing dbt CLI invocation to fail when
materializing assets was fixed. This issue occurred only for assets for
which `OpExecutionContext` was used as the type hint for the context.