[dagster-databricks] Skip tests reliant on external infrastructure (#29824)
## Summary & Motivation

Addresses https://github.com/dagster-io/dagster/issues/29526, by adding
a pytest `skipif` condition on the presence of the `DATABRICKS_HOST`
environment variable. It appears these tests were built assuming AWS
infrastructure, so some basic parsing of the URL was done to set the
`node_type_id` to Azure equivalents, with a fallback to the previous
default.

Some basic cleaning for typos and unused code was done elsewhere in the
module.

## How I Tested These Changes

Locally targeting the `dagster-databricks` library path.