[dagster-airbyte] Raise custom exception on conflicting group_name in Airbyte asset decorator (#29976)
## Summary & Motivation

Following changes made in
https://github.com/dagster-io/dagster/pull/29968.

~~This PR sets the `group_name` using
`replace_attributes(group_name=...)` in the Airbyte asset decorator
instead of passing the group name to the multi-asset definitions. This
allows users to either set the `group_name` at the spec level in the
translator or the asset decorator level without raising an exception.~~

~~Note that if a `group_name` is passed to the asset decorator, it
overrides the group name set in the spec.~~

Edit:

After discussion with @benpankow, we will explicitly disallow setting a
group name on the asset decorator and in the translator by raising a
custom exception.

## How I Tested These Changes

Additional tests with BK