[docs] extend note for dbt defer feature (#30034)
## Summary & Motivation
To use this feature the `dagster-cloud ci dagster-dbt project
manage-state command` also needs to be run in prod deployment before it
can be run from in a branch deployment. So a step in the CI/CD workflow
with it must be added for the prod deployment and deployment executed
with it.

## How I Tested These Changes
👀

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>
Co-authored-by: Nikki Everett <neverett@users.noreply.github.com>