[bug] Fix issue with op_tags being dropped for observable source assets (#30088)
## Summary & Motivation

Reported by a user -- we were not properly passing this information through. Originally, I fixed this by just forwarding the tags through this call, but then I realized there were a lot of other properties we were dropping at the same time.

Switched to using dagster_internal_init instead, and filled in a bunch of other properties (most of which will probably always be None but still)


## How I Tested These Changes

## Changelog

Fixed an issue that would cause `op_tags` set on `@observable_source_asset`s to be dropped from the underlying step context object when executed.