[docs] Reference architecture tags (#30091)
## Summary & Motivation
Adding `tags` to examples for reference architectures vs code examples.
Also updating `docs/src/theme/DocCard/index.tsx` to include an optional
parameter for reference architectures so the tag can appear on the card.

![Screenshot 2025-05-19 at 12 59
41 PM](https://github.com/user-attachments/assets/2a7a32d4-5a7a-4bb0-a88d-4201d5b56cda)

I don't love that it is inconsistent in one of the three having a line
break. Can't think of how to fix without changing to a single word.

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.

---------

Co-authored-by: Nikki Everett <neverett@users.noreply.github.com>