Remove strict typer pin
Summary:
Sending this out for discussion - if dagster-dg is going to be installable in proejct venvs and not just globally, this pin will likely be too restrictive. Is there more information on these private typer APIs that we are using somehwere? For now I changed this to `<0.16.0`.

Also pre-emptively added a watchdog pin since they are unafraid to release breaking changes.