[Asset Graph] Persist zoom level to localStorage (#30128)
## Summary & Motivation

Persist zoom level to localStorage so that when you visit a graph it
restores your previous zoom level.


https://linear.app/dagster-labs/issue/OPER-2016/persist-users-zoom-level-on-the-asset-graph


https://github.com/user-attachments/assets/4125f58c-b4d5-44e1-afab-fe4a197dde83

The setting is shared across all asset-graphs and the zoom appropriate
for one might not be appropriate for another but I think this edge case
isn't a big deal and it's a better experience overall to preserve it.