[dg] [ontology] Rename lib folder to components (BUILD-1196) (#30053)
## Summary & Motivation

Rename the "lib" folder in scaffolded projects to "components". This
does not require any backcompat since the "lib" value is not assumed in
our code, but sourced from the entry point declaration in a project.

## How I Tested These Changes

Existing test suite.

## Changelog

The "lib" folder in newly scaffolded projects has been renamed to
"components".