Optionally truncate cloud compute logs
We have a longstanding issue with our cloud compute logs where we
rewrite the same key over and over again as the file grows because each
time we upload, we upload the entire locally captured file.

We should address that - presumably by either mirroring stdout/stderr to
chunked files or by writing some sort of chunk reader that can iterate
over the full local file.

This stops a step short by providing an env var that can be set to
truncate the max file size to upload. The implementation is pretty
naive. It maintains current behavior up until the limit. Once the local
file is larger than our truncation limit, it writes a truncated temp
file, uploads that one, and sets a flat to no longer upload anything
else.