[docs][da] Rework Declarative Automation documentation (#29931)
## Summary & Motivation

This is a large set of changes to the Declarative Automation documentation.

A quick summary:

### Reworking the landing page

The previous landing page had a lot of information that was presented in a pretty overwhelming way. This pares down that information to the basics, and heavily guides users towards the builtin automation conditions, putting these in tabs to make it easy to compare and contrast the different options

### Add more precise behavior descriptions for each builtin type

This is the "Behavior" section in each of the tabs. We had "loose" descriptions of the basic thing that each condition was meant to do, but were lacking a more precise technical explanation for the exact behavior that each condition would have.

### Separate customization pages for each builtin condition type

The previous customization page was already somewhat hard to navigate, even with a fairly limited set of examples. Given that we're guiding users to use builtin conditions and treating those as top-level concepts, having separate customization pages for each that focus on the types of customizations that make sense

### More example customizations

Added more examples of customizing each of the individual condition types

### Create a standalone page for describing AutomationConditionSensorDefinitions

This was squished into the overview page, which wasn't ideal, as you don't need this information when you're just getting started. Moving this out into its own separate page also gave me room to add some examples for how to customize behavior of these sensors.

After this PR, I have the following followups:

- Add specific information about setting automation conditions on asset checks
- Update automation landing page to have a description of when you should use DA vs. other automation methods
- Add "troubleshooting" guide