[Catalog] Add `isFullScreenEnabled` atom to control whether a view is full page screenable (#30120)
## Summary & Motivation
"Full screen" mode causes the page's navigation to disappear. It is
intended for use by the lineage page in the catalog.

Currently if you turn fullscreen mode on, and navigate away by clicking
a link from the asset graph's right sidebar (eg. "View in catalog" link
for an asset) then fullscreen mode stays on and there's no way to turn
it off to see the page's navigation. To fix this I'm adding a second
atom `isFullScreenEnabled` and a hook `useFullScreenEnabledView`

## How I Tested These Changes


https://github.com/user-attachments/assets/cc2962fd-4ac4-4159-b7d6-20626a8b1937