remote populate-cache argument from scaffold_project (#30169)
## Summary & Motivation
Now that populating the cache happens in process rather than over IPC, it is not safe to do this when scaffolding a project - it will populate the cache with the plugins in the global environment rather than the plugins in the venv.

Incidentally the same thing happens if you accidentally run dg list plugins in the wrong venv now while in a project. Do we need to do anything more proactive to detect and prevent that?

## How I Tested These Changes
BK