[Asset Graph] Hide non SDAs again (#30204)
## Summary & Motivation

in https://github.com/dagster-io/dagster/pull/30179 I included a change
to back `useAssetGraphData` by `useAllAssets` which includes external
assets. This was a change in behavior and led to asset graphs including
non SDAs. Instead we should go back to relying on the
`options.externalAssets` prop so that we only include external assets in
the asset selection input (which uses this hook and passes this prop) .

## How I Tested These Changes

Checked that asset graphs no longer include external assets