[dg] Add classnames as aliases to scaffold defs subcommands (BUILD-1225) (#30219)
## Summary & Motivation

This adds class names as command aliases for `dg scaffold defs`
subcommands when those class names are unique. For example, if I have a
component `some_lib.FooComponent`, and that is the only component named
`FooComponent`, then I can just do:

```
$ dg scaffold defs FooComponent ...
```

And it will work as if I typed the fully-qualified name. However, if
`some_other_lib.FooComponent` is also in scope, the alias is not added
and the other above command will fall back to the substring match menu.

## How I Tested These Changes

New unit tests.

## Changelog

When component classnames are unique, you can now use just the classname
as an alias for the fully qualified name when running `dg scaffold defs
<component>`.

---------

Co-authored-by: Nick Schrock <schrockn@dagsterlabs.com>