Make the default dagster_dg.plugin dagster_dg_cli.plugin instead (#30284)
## Summary & Motivation
This is the last place we reference dagster_dg. I'm not strongly wed to
this, but if we're being consistent this should be called dagster_dg_cli
since that's the new name of the package.

I left the old dagster_dg.plugin as an option, and I don't make it warn
if you set it like you do if you set dagster_dg.library.

## How I Tested These Changes
Adjusted test cases