[dg] Add support for autodefs_module_name in workspace loading (#30133)
Adds a new `CodePointer` and associated cli options for targeting an
"autodefs module" which will do a components root load at that target.
This removes the need for a manual `definitions.py`

Moving to this is done via a new tool section entry `autodefs_module`
which takes the module to start crawling at.


## How I Tested These Changes

removed the root definitions.py in dagster open platform and added the
`autoload` value to `pyproject.toml`