[ui] ListItem padding customizability (#30330)
## Summary & Motivation

Make the padding for the new `ListItem` component customizable in the same manner  as `Box`. This is to support being able to use these items in page-size use cases, as well as smaller ones that need smaller padding.

I liked this better than trying to define an enum like `small`, `medium`, etc. If we decide we want to be more constrained about that, we can come back to it.

<img width="428" alt="Screenshot 2025-05-29 at 15 22 08" src="https://github.com/user-attachments/assets/0b060155-3894-4a83-a18b-8b8a0ffc8ba9" />


## How I Tested These Changes

Storybook