Add resolved_ variants for all getters on Definitions (#30369)
## Summary & Motivation

As discussed in https://github.com/dagster-io/dagster/pull/30250 we are going to do a series of changes to make Definitions more usable as a "dumb" object suitable for testing and existence in an intermediate state. This has increased salience in a world with a `defs` hierarchy, where Definitions are created frequently, merged, and discarded without resolution (resolution defined as the creation of the underlying `RepositoryDefinition`). Additionally defs paths are independently loadable but not resolved.

This first change introduces `resolve_` variants of all getters that required a fully resolved Definitions object. This convention will make it clear what operaates cause resolution versus not.

## How I Tested These Changes

BK