Refactor map_resolved_asset_specs to eliminate boolean flag (#30370)
## Summary & Motivation

In certain code paths (that are constrained in certain ways) we allow SourceAssets and CacheableAssetsDefinitions to be mapped. This refactors the inner function that allows that to not use a boolean argument. To use a fancy term this reduces the cyclomatic complexity of the code path, making it easier to understand. 

## How I Tested These Changes

BK