Eliminate global component load context (#30360)
## Summary & Motivation

While writing testing utilities and other code I've noted that we have a global context variable to store a ComponentLoadContext. I think this is easier to screw up and unnecessary. The system will be better, more testable, and more modular if we can get away with not having it.

Two things in order to accomplish this

* If you need access to the context, you have to do in a `@definitions` callback or a `build_defs`. I think the users will thank us for pushing more compute out of the import path
* In another case, I added a stash dictionary to the `ResolutionContext`. I think this is the right thing to do, as it solidfies as ResolutionContext as "lower" in the stack than Components. The component load context is only retrieved from the stash when we know we are dealing with components code. I still want to preserve the option value to use Resolved outside of components contexts.

This resolves is a net negative LOC PR, which is good, and the lines we remove are easy ones to miss. Wouldn't be surprised if we already miss callsites.

## How I Tested These Changes

BK