[asset health] handle computing health data for assets with no definition (#30319)
## Summary & Motivation
Some assets have no node on the asset graph (for example, you can report
an asset materialization for a key that otherwise has no references in
the asset graph, see the new examples added to the toys repo to see an
example)

Computing the health for an asset sometimes required getting the node
for an asset from the asset graph, this would fail for the kinds of
assets described above since there is no node in the asset graph. This
PR fixes that by adding a different path for each health component if
the asset has no node in the graph. For materialization, we are able to
compute some information based on the record in the DB (this falls short
if the reported materializations have partitions associated with them,
but we don't handle this perfectly in other parts of the UI right now,
so i think the current behavior is fine as a starting place).

For asset checks and freshness, we return NOT_APPLICABLE results if the
asset has no definition. Technically you can report check results for an
asset that has no definition, but we don't have storage methods to get
these check results (you can only fetch results if you have the check
key, and you can only get the check keys for an asset via the asset
graph), so we don't display them anywhere in the UI right now. A future
improvement would be to add these storage methods to allow us to display
the check results throughout the UI and compute health based on them.

If streamline is enabled, we are able to report on the check events that
are reported for assets that don't have nodes in the asset graph since
we can build up the data needed to report health as the events come in
and then fetch that data by asset key, rather than needing the check
keys


In this screenshot `no_def_observable` and `no_def_materializable` are
assets with no definition, but there are jobs the emit
observation/materialization events for assets with this these keys. We
are now able to show results for them in the UI
<img width="1900" alt="Screenshot 2025-05-29 at 10 24 01 AM"
src="https://github.com/user-attachments/assets/727f54cf-0511-46a1-abea-c78ba5405cd5"
/>

Also updates the UI so that we show the health status report card on the
asset details page for all assets, even if they don't have a definition
<img width="767" alt="Screenshot 2025-05-29 at 11 55 54 AM"
src="https://github.com/user-attachments/assets/252643b7-0982-4c8e-93b6-8b2a188dd71a"
/>

## How I Tested These Changes
new unit tests in https://github.com/dagster-io/internal/pull/16021